requirements.txt
pip==22.0.4
asgiref==3.5.0
Django==3.2
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
django-allauth==0.45.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
dj-rest-auth==2.2.5
requests==2.25.1
django-filter==22.1